Refactored os walking loop and renamed pathanme to filename.
While iterating through the files the variable pathname was used as
a filename though the variable filename was present from the loop before.
This resulted in a small problem in which pattern without a slash did not
not match as the wrong variable was used.

To simplify everything I renamed the variable pathname, which was used as
filename anyway, to filename.